split core into several sub services
